pyproject.toml
[tool.poetry]
name = "editora-v2-movie-maker"
version = "0.1.0"
description = "Video engine for Editora"
authors = ["Kishan Jayaraman <kishan.jay@editora.ai>", "Victor Souto | Plank <souto.victor@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pydantic = {extras = ["email"], version = "^2.10.6"}
loguru = "^0.7.2"
openai = "^1.55.1"
python-dotenv = "^1.0.1"
rich = "^13.9.4"
pyyaml = "^6.0.2"
firebase-admin = "^6.6.0"
wget = "^3.2"
fastapi = "^0.115.5"
python-multipart = "^0.0.6"
aiofiles = "^24.1.0"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
google-cloud-secret-manager = "^2.21.1"
elevenlabs = "^1.13.0"
assemblyai = "^0.35.1"
moviepy = "^2.1.1"
pillow-avif-plugin = "^1.4.6"
google-cloud-vision = "^3.8.1"
google-cloud-videointelligence = "^2.13.5"
google-cloud-aiplatform = "^1.74.0"
grpcio = "1.67.1"
google-cloud-logging = "^3.11.3"
mailchimp-transactional = "^1.0.56"
pillow-heif = "^0.21.0"
pysrt = "^1.1.2"
pydub = "^0.25.1"
stytch = "^12.7.0"
pyjwt = "^2.10.1"
bs4 = "^0.0.2"
jmespath = "^1.0.1"
redfin = "^0.1.1"
inflection = "^0.5.1"
scrapfly-sdk = "^0.8.21"
parsel = "^1.10.0"
usaddress = "^0.5.13"
playwright = "^1.50.0"
opencv-python-headless = "^4.10.0"
numpy = "^1.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
